finished reload method
